[#176443163] move saga folder from store folder
